requirements.txt
python-telegram-bot==20.7
beautifulsoup4==4.12.2
